requirements.txt
pybit==5.8.0
pandas==2.0.3
ta==0.10.2
openai==1.3.0
python-dotenv==1.0.0
qdrant-client==1.9.0
flask==2.3.0